requirements.txt
# Core dependencies
asyncio
aiohttp>=3.9.0
aiofiles>=23.2.1
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Web scraping and crawling
scrapy>=2.11.0
selenium>=4.15.0
beautifulsoup4>=4.12.2
requests>=2.31.0
fake-useragent>=1.4.0

# Search engine APIs
google-api-python-client>=2.110.0
python-bing-search>=0.3.0
googlesearch-python>=1.2.3

# Image processing and recognition
opencv-python>=4.8.1.78
opencv-contrib-python>=4.8.1.78
face-recognition>=1.3.0
Pillow>=10.1.0
imagehash>=4.3.1
numpy>=1.24.3

# Machine learning for content matching
scikit-image>=0.22.0
tensorflow>=2.15.0
torch>=2.1.0
torchvision>=0.16.0

# Task queue and scheduling
celery>=5.3.4
redis>=5.0.1
APScheduler>=3.10.4

# Proxy support
requests[socks]>=2.31.0
aiohttp-socks>=0.8.4

# Rate limiting
aiohttp-ratelimiter>=0.3.0
backoff>=2.2.1

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Monitoring and logging
structlog>=23.2.0
sentry-sdk>=1.38.0
prometheus-client>=0.19.0

# Email and notifications
sendgrid>=6.10.0
jinja2>=3.1.2

# Utilities
pydantic>=2.5.0
typer>=0.9.0
rich>=13.7.0
httpx>=0.25.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0